requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==20.3.0
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.4
distlib==0.3.1
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
hypothesis==6.1.1
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyJWT==2.0.1
pylint==2.6.0
pylint-venv==2.1.1
pyparsing==2.4.7
pytest==6.2.2
requests==2.25.1
six==1.15.0
sortedcontainers==2.3.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
mypy==0.800
pycodestyle==2.7.0
simplejson==3.17.2
flask-mail==0.9.1
pytz==2020.4
cryptography==3.2.1
emojis==0.6.0
pytest-timeout==1.4.2
pydictionary==2.0.1
mypy==0.800
english-words==1.0.3
